imDefIc: Clear fabricated state on unfocus.
When unsetting focus, the event filter is removed. This means a pending
fabricated event may not yet be sent to filter.

All the fabricated event state should be cleared and the pending sync
reply sent back as if the state is unfabricated.

Fix #235

Part-of: <https://gitlab.freedesktop.org/xorg/lib/libx11/-/merge_requests/283>