xkb: Fix default key types
Before this commit, using xkbcomp to compile a keymap that omits the key
type `ALPHABETIC` leads to either a wrong serialization (`-xkb` output)
or a server error such as:

    X Error of failed request:  BadValue (integer parameter out of range for operation)
      Major opcode of failed request:  135 (XKEYBOARD)
      Minor opcode of failed request:  9 (XkbSetMap)
      Value in failed request:  0x8010202
      Serial number of failed request:  55
      Current serial number in output stream:  60

Indeed, there is an error in the default key maps: the second field of
`XkbKTMapEntryRec` if a level index, not a modifier mask. While it worked
*by chance* for `ShiftMask` (e.g. the `TWO_LEVEL` type), it does not work
for `LockMask` (e.g. the `ALPHABETIC` type).

Fixed by using the correct level indices.

It probably went unnoticed for so long because xkbcomp and libxkbcommon
< 1.12 always use all the key types, while libxkbcommon ≥ 1.12 does not
serialize unused key types. The latter case may result in a keymap that
omits `ALPHABETIC`, e.g. with keyboard layouts with 4+ levels which use
alternative key types such as `FOUR_LEVEL_ALPHABETIC`.

Signed-off-by: Pierre Le Marre <dev@wismill.eu>
Part-of: <https://gitlab.freedesktop.org/xorg/lib/libx11/-/merge_requests/292>